test(admin-phase3): add unit and feature tests for Phase 3 features
Unit tests:
- FailedJobTest: job_name accessor (namespaced/unknown), timestamps flag, table name
- SystemHealthPageTest: application metrics (pending/active/online counts), reverb channel
  parsing, graceful empty return on HTTP failure and connection exception
- DriverQueuePageTest: FIFO ordering, queue filtering, cooldown counting, avg wait as int

Feature tests:
- AdminLiveUpdateEventTest: channel name, broadcastAs, broadcastWith structure
- AdminListenersTest: all 8 listeners fire AdminLiveUpdate with correct type/payload;
  BroadcastAdminDriverRejoined only fires on completed/cancelled, not active statuses,
  not when driver_id is null; EventServiceProvider wiring verified
- AdminChannelAuthorizationTest: admin.live channel allows admin, blocks rider/driver/both

Also fixes: FailedJob adds $guarded=[], DriverQueuePage casts avg_wait_minutes to int

+++ backend/app/Filament/Pages/DriverQueuePage.php
@@ -38,7 +38,7 @@ class DriverQueuePage extends Page
 
         return [
             'total' => $drivers->count(),
-            'avg_wait_minutes' => round($avgWaitMinutes),
+            'avg_wait_minutes' => (int) round($avgWaitMinutes),
             'in_cooldown' => $inCooldown,
         ];
     }
